LayoutTests/http/tests/security/dataURL/xss-DENIED-from-data-url-sub-frame-2-level-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-from-data-url-sub-frame-2-level.html from frame with URL data:text/html,<html><head><script>function test() {try {top.document.getElementById("accessMe").innerHTML = "FAIL: Cross frame access from a data: URL inside another data: URL was allowed.";} catch (e) {}if (window.layoutTestController)layoutTestController.notifyDone();}</script></head><body onload="test();"><p>Inner-inner iframe.</p></body></html>. Domains, protocols and ports must match.

This tests that a data: URL loaded in an iframe inside another data: URL loaded iframe doesn't have access to the main frame.


PASS: Cross frame access from a data: URL inside another data: URL was denied.

--------
Frame: 'aFrame'
--------
Inner iframe.



--------
Frame: '<!--framePath //aFrame/<!--frame0-->-->'
--------
Inner-inner iframe.
